feat:  optimize navbar component and reused it in the menu for small devices. Change the name from fullscreen-menu to drawer-menu
